requirements-dev.txt
-r requirements.txt

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
moto[s3,dynamodb]==5.0.0
boto3-stubs[s3,dynamodb,lambda,apigateway]==1.34.34

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Development
ipython==8.21.0
